requirements.txt
# clean code
#black
#corner
#flake8
#flake8-bugbear
#isort
#pre-commit
#pycodestyle

argparse

bilby
deepdish
gwpy
hurry.filesize
lalsuite
matplotlib
numpy
opencv-python-headless
pandas
plotly
scipy
